backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
google-cloud-speech==2.28.0
google-cloud-translate==3.18.0
redis==5.2.1
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
hypothesis==6.122.3